Use HTTP POST when deleting attachments
This fixes deleting attachments on tomcat where HTTP DELETE (and PUT)
are disabled by default. Enabling them is also not possible because that
would allow users to delete attachments if they know the id of the
attachment on the file-system (for example with curl, by running
curl -X DELETE http://mamute.local/attachments/1).